fix: credential table moved to x/service
