+++ proto/sonrhq/sonr/identity/module/v1/state_query.proto
@@ -23,16 +23,6 @@ service StateQueryService {
   rpc GetBlockchainByName(GetBlockchainByNameRequest) returns (GetBlockchainByNameResponse) {}
   // ListBlockchain queries the Blockchain table using prefix and range queries against defined indexes.
   rpc ListBlockchain(ListBlockchainRequest) returns (ListBlockchainResponse) {}
-  // Get queries the Credential table by its primary key.
-  rpc GetCredential(GetCredentialRequest) returns (GetCredentialResponse) {}
-  // GetCredentialByOriginHandle queries the Credential table by its OriginHandle index
-  rpc GetCredentialByOriginHandle(GetCredentialByOriginHandleRequest) returns (GetCredentialByOriginHandleResponse) {}
-  // GetCredentialByCredentialId queries the Credential table by its CredentialId index
-  rpc GetCredentialByCredentialId(GetCredentialByCredentialIdRequest) returns (GetCredentialByCredentialIdResponse) {}
-  // GetCredentialByPublicKey queries the Credential table by its PublicKey index
-  rpc GetCredentialByPublicKey(GetCredentialByPublicKeyRequest) returns (GetCredentialByPublicKeyResponse) {}
-  // ListCredential queries the Credential table using prefix and range queries against defined indexes.
-  rpc ListCredential(ListCredentialRequest) returns (ListCredentialResponse) {}
   // Get queries the Identifier table by its primary key.
   rpc GetIdentifier(GetIdentifierRequest) returns (GetIdentifierResponse) {}
   // ListIdentifier queries the Identifier table using prefix and range queries against defined indexes.
@@ -225,131 +215,6 @@ message ListBlockchainRequest {
 message ListBlockchainResponse {
   // values are the results of the query.
   repeated Blockchain values = 1;
-  // pagination is the pagination response.
-  cosmos.base.query.v1beta1.PageResponse pagination = 2;
-}
-
-// GetCredentialRequest is the StateQuery/GetCredentialRequest request type.
-message GetCredentialRequest {
-  // id specifies the value of the id field in the primary key.
-  uint64 id = 1;
-}
-
-// GetCredentialResponse is the StateQuery/GetCredentialResponse response type.
-message GetCredentialResponse {
-  // value is the response value.
-  Credential value = 1;
-}
-
-// GetCredentialByOriginHandleRequest is the StateQuery/GetCredentialByOriginHandleRequest request type.
-message GetCredentialByOriginHandleRequest {
-  string origin = 1;
-  string handle = 2;
-}
-
-// GetCredentialByOriginHandleResponse is the StateQuery/GetCredentialByOriginHandleResponse response type.
-message GetCredentialByOriginHandleResponse {
-  Credential value = 1;
-}
-
-// GetCredentialByCredentialIdRequest is the StateQuery/GetCredentialByCredentialIdRequest request type.
-message GetCredentialByCredentialIdRequest {
-  bytes credential_id = 1;
-}
-
-// GetCredentialByCredentialIdResponse is the StateQuery/GetCredentialByCredentialIdResponse response type.
-message GetCredentialByCredentialIdResponse {
-  Credential value = 1;
-}
-
-// GetCredentialByPublicKeyRequest is the StateQuery/GetCredentialByPublicKeyRequest request type.
-message GetCredentialByPublicKeyRequest {
-  bytes public_key = 1;
-}
-
-// GetCredentialByPublicKeyResponse is the StateQuery/GetCredentialByPublicKeyResponse response type.
-message GetCredentialByPublicKeyResponse {
-  Credential value = 1;
-}
-
-// ListCredentialRequest is the StateQuery/ListCredentialRequest request type.
-message ListCredentialRequest {
-  // IndexKey specifies the value of an index key to use in prefix and range queries.
-  message IndexKey {
-    // key specifies the index key value.
-    oneof key {
-      // id specifies the value of the Id index key to use in the query.
-      Id id = 1;
-      // handle specifies the value of the Handle index key to use in the query.
-      Handle handle = 2;
-      // origin_handle specifies the value of the OriginHandle index key to use in the query.
-      OriginHandle origin_handle = 3;
-      // credential_id specifies the value of the CredentialId index key to use in the query.
-      CredentialId credential_id = 4;
-      // public_key specifies the value of the PublicKey index key to use in the query.
-      PublicKey public_key = 5;
-    }
-    
-    message Id {
-      // id is the value of the id field in the index.
-      // It can be omitted to query for all valid values of that field in this segment of the index.
-      optional uint64 id = 1;
-    }
-    
-    message Handle {
-      // handle is the value of the handle field in the index.
-      // It can be omitted to query for all valid values of that field in this segment of the index.
-      optional string handle = 1;
-    }
-    
-    message OriginHandle {
-      // origin is the value of the origin field in the index.
-      // It can be omitted to query for all valid values of that field in this segment of the index.
-      optional string origin = 1;
-      // handle is the value of the handle field in the index.
-      // It can be omitted to query for all valid values of that field in this segment of the index.
-      optional string handle = 2;
-    }
-    
-    message CredentialId {
-      // credential_id is the value of the credential_id field in the index.
-      // It can be omitted to query for all valid values of that field in this segment of the index.
-      optional bytes credential_id = 1;
-    }
-    
-    message PublicKey {
-      // public_key is the value of the public_key field in the index.
-      // It can be omitted to query for all valid values of that field in this segment of the index.
-      optional bytes public_key = 1;
-    }
-  }
-  
-  // query specifies the type of query - either a prefix or range query.
-  oneof query {
-    // prefix_query specifies the index key value to use for the prefix query.
-    IndexKey prefix_query = 1;
-    // range_query specifies the index key from/to values to use for the range query.
-    RangeQuery range_query = 2;
-  }
-  // pagination specifies optional pagination parameters.
-  cosmos.base.query.v1beta1.PageRequest pagination = 3;
-  
-  // RangeQuery specifies the from/to index keys for a range query.
-  message RangeQuery {
-    // from is the index key to use for the start of the range query.
-    // To query from the start of an index, specify an index key for that index with empty values.
-    IndexKey from = 1;
-    // to is the index key to use for the end of the range query.
-    // The index key type MUST be the same as the index key type used for from.
-    // To query from to the end of an index it can be omitted.
-    IndexKey to = 2;
-  }
-}
-
-// ListCredentialResponse is the StateQuery/ListCredentialResponse response type.
-message ListCredentialResponse {
-  // values are the results of the query.
-  repeated Credential values = 1;
   // pagination is the pagination response.
   cosmos.base.query.v1beta1.PageResponse pagination = 2;
 }
